pyproject.toml
[project]
name = "langgraph-fun"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "ddgs==9.5.5",
    "distro==1.9.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-ws==0.7.2",
    "idna==3.10",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain-core==0.3.75",
    "langgraph==0.6.7",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.6.4",
    "langgraph-sdk==0.2.6",
    "langsmith==0.4.27",
    "lmstudio==1.5.0",
    "lxml==6.0.1",
    "msgspec==0.19.0",
    "openai==1.107.0",
    "orjson==3.11.3",
    "ormsgpack==1.10.0",
    "packaging==25.0",
    "primp==0.15.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyyaml==6.0.2",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "sniffio==1.3.1",
    "tenacity==9.1.2",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "wsproto==1.2.0",
    "xxhash==3.5.0",
    "zstandard==0.24.0",
]

[dependency-groups]
dev = [
    "black>=23.0.0",
    "coverage>=7.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-xdist>=3.0.0",
]
